# openapi.yaml
openapi: 3.0.3
info:
  title: API MiniBlog M2
  version: 1.0.0
  description: API REST con Node.js, Express y PostgreSQL para gestionar authors, posts y comments.
  contact:
    name: Jazmin Morinigo
    email: noejazminmrivas@gmail.com
servers:
  - url: http://localhost:3000
    description: Servidor local
  - url:
...
paths:
...
  /authors:
    get:
      summary: Listar authors
      responses:
        "200":
          description: Lista de authors
    post:
      summary: Crear author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorInput"
      responses:
        "201":
          description: Author creado
        "400":
          description: Datos invalidos
        "409":
          description: Email duplicado
...
  /posts:
    get:
      summary: Listar posts
      responses:
        "200":
          description: Lista de posts
    post:
      summary: Crear post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post creado
        "400":
          description: Datos invalidos
        "404":
          description: Author no encontrado
...
  /comments:
    get:
      summary: Listar comments
      responses:
        "200":
          description: Lista de comments
    post:
      summary: Crear comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment creado
        "400":
          description: Datos invalidos
        "404":
          description: Post o author no encontrado
...
components:
...
  schemas:
    AuthorInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: Ana Garcia
        email:
          type: string
          example: ana@example.com
        bio:
          type: string
          example: Autora de contenidos backend.
    PostInput:
      type: object
      required:
        - title
        - content
        - author_id
      properties:
        title:
          type: string
          example: Introduccion a Node.js
        content:
          type: string
          example: Node.js permite ejecutar JavaScript en el servidor.
        author_id:
          type: integer
          example: 1
        published:
          type: boolean
          example: true
    CommentInput:
      type: object
      required:
        - content
        - post_id
        - author_id
      properties:
        content:
          type: string
          example: Muy buen post.
        post_id:
          type: integer
          example: 1
        author_id:
          type: integer
          example: 1